Moved new scheduling tests into proper folders
