+++ autoscheduler/frontend/src/tests/ui/SchedulePreview.test.tsx
@@ -2,10 +2,10 @@ import * as React from 'react';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 import { render, fireEvent } from '@testing-library/react';
-import SchedulePreview from '../components/SchedulingPage/SchedulePreview/SchedulePreview';
-import autoSchedulerReducer from '../redux/reducer';
-import { replaceSchedules } from '../redux/actions/schedules';
-import { testSchedule1, testSchedule2 } from './testSchedules';
+import SchedulePreview from '../../components/SchedulingPage/SchedulePreview/SchedulePreview';
+import autoSchedulerReducer from '../../redux/reducer';
+import { replaceSchedules } from '../../redux/actions/schedules';
+import { testSchedule1, testSchedule2 } from '../testSchedules';
 
 describe('SchedulePreview component', () => {
   describe('updates the selected schedule', () => {
